Add a db config migration for old mediacore url setting

+++ local/mediacore/db/upgrade.php
@@ -31,8 +31,27 @@
  */
 
 function xmldb_local_mediacore_upgrade($oldversion) {
-    global $CFG, $DB;
-    $DB->delete_records('config_plugins',
-            array('plugin' => 'local_mediacore_courselti'));
+    global $DB;
+
+    $old_record = $DB->get_record('config_plugins',
+            array('plugin'=>'local_mediacore', 'name'=>'url'));
+
+    // Add the host value to the db and delete the old url value
+    if ($old_record) {
+        $hostname = parse_url($old_record->value, PHP_URL_HOST);
+        $port = parse_url($old_record->value, PHP_URL_PORT);
+        $host = $hostname;
+        if ($port) {
+            $host .= ':' . $port;
+        }
+        $new_record = new stdClass();
+        $new_record->plugin = 'local_mediacore';
+        $new_record->name = 'host';
+        $new_record->value = $host;
+
+        $DB->insert_record('config_plugins', $new_record, false);
+        $DB->delete_records('config_plugins',
+                array('plugin' => 'local_mediacore', 'name' => 'url'));
+    }
     return true;
 }
